fix: create an atom & refactor
